+++ db/Dockerfile
@@ -1,0 +1,3 @@
+FROM postgres
+ADD init.sql /docker-entrypoint-initdb.d
+EXPOSE 5432
